refactor : 하드코딩 값 상수화 및 테스트 로직 패키지 변경
