+++ src/main/java/stringaddcalculator/domain/StringAddCalculator.java
@@ -8,20 +8,24 @@ public class StringAddCalculator {
 
     private static final String DELIMITER = ":|,";
     private static final String CUSTOM_DELIMITER = "//(.)\n(.*)";
+    public static final int ZERO_NUMBER = 0;
+    private static final int CUSTOM_DELIMITER_CHERKER_ONE = 1;
+    private static final int CUSTOM_DELIMITER_CHECKER_TWO = 2;
+    private static Matcher expressionMatcher;
 
     public static int sum(String expression) {
         if (ValidationNumber.isBlank(expression)) {
-            return 0;
+            return ZERO_NUMBER;
         }
 
         String[] numbers = splitDelimiterCheck(expression);
 
         return Arrays.stream(numbers)
                 .map(number -> toInt(number))
-                .reduce(0, Integer::sum);
+                .reduce(ZERO_NUMBER, Integer::sum);
     }
 
-    private static String[] expressionSplit(String expression) {
+    private static String[] splitExpression(String expression) {
         return expression.split(DELIMITER);
     }
 
@@ -32,12 +36,12 @@ public class StringAddCalculator {
     }
 
     public static String[] splitDelimiterCheck(String expression) {
-        Matcher expressionMatcher = Pattern.compile(CUSTOM_DELIMITER).matcher(expression);
+        expressionMatcher = Pattern.compile(CUSTOM_DELIMITER).matcher(expression);
         if (expressionMatcher.find()) {
-            String customDelimiter = expressionMatcher.group(1);
-            return expressionMatcher.group(2).split(customDelimiter);
+            String customDelimiter = expressionMatcher.group(CUSTOM_DELIMITER_CHERKER_ONE);
+            return expressionMatcher.group(CUSTOM_DELIMITER_CHECKER_TWO).split(customDelimiter);
         }
-        return expressionSplit(expression);
+        return splitExpression(expression);
     }
 
 }
